moved maybeIdentifierUnderPosition under activeNode
