+++ src/inspection/activeNode/activeNode.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { NodeIdMap } from "../../parser";
+import { Option } from "../../common";
+import { Ast, NodeIdMap } from "../../parser";
 import { Position } from "../position";
 
 // An ActiveNode represents the context a text editor user expects their cursor to be in.
@@ -15,4 +16,7 @@ export interface ActiveNode {
     // A full parental ancestry of the root.
     // [root, parent of root, parent of parent of root, ...]
     readonly ancestry: ReadonlyArray<NodeIdMap.TXorNode>;
+    // A cache of an evaluation that otherwise would need to be reevaluated.
+    // If ActiveNode's leaf is an identifier then store the indirection to it as an Ast node.
+    readonly maybeIdentifierUnderPosition: Option<Ast.Identifier | Ast.GeneralizedIdentifier>;
 }
